permisos por roles

+++ views/batch/batch.php
@@ -104,10 +104,7 @@ require_once '../app/controllers/batch/show_events.php';
 
             <!-- checkbox -->
             <div class="input">
-
-
                 <label for=""> Integración Disponible </label>
-
                 <div class="checkboxes">
 
                     <div class="check">
@@ -122,7 +119,7 @@ require_once '../app/controllers/batch/show_events.php';
                                     )
                                     : ''
 
-                                ?> type="checkbox" name="availabity" id="si" onclick="handleCheckBox('si')">
+                                ?> type="checkbox" name="availabity" id="si" onclick="handleCheckBox('si')" <?php echo $_SESSION['role'] == 'reader' ? 'disabled' : '' ?>>
 
 
                     </div>
@@ -137,67 +134,74 @@ require_once '../app/controllers/batch/show_events.php';
                                         : ''
                                     )
                                     : ''
-                                ?> type="checkbox" name="inavailabity" id="no" onclick="handleCheckBox('no')">
+                                ?> type="checkbox" name="inavailabity" id="no" onclick="handleCheckBox('no')" <?php echo $_SESSION['role'] == 'reader' ? 'disabled' : '' ?>>
 
                     </div>
 
                 </div>
-
-
-
             </div>
-
-
 
             <!-- Hora de ejecucion -->
             <div class="input">
 
                 <label for=""> Hora de ejecución </label>
-                <input value="<?php echo isset($execution_schedule) ? $execution_schedule : ''; ?>" type="time" id="execution_schedule" name="execution_schedule" required />
-
+                <input value="<?php echo isset($execution_schedule) ? $execution_schedule : ''; ?>" type="time" id="execution_schedule" name="execution_schedule" required <?php echo $_SESSION['role'] == 'reader' ? 'disabled' : '' ?> />
 
             </div>
-            <input type="submit" name="batch-submit" value="Guardar configuracion">
+            <?php
+            if ($_SESSION['role'] == 'admin') {
+                echo "
+                    <input type='submit' name='batch-submit' value='Guardar configuracion'>
+                    ";
+            }
+            ?>
+
         </form>
 
         <div class="form-target">
 
 
             <div class="new-noti-target">
-                <h3>Nuevo destinatarios</h3>
-
-                <form class="input-new-target" method="post">
-
-                    <input type="hidden" name="id_batch" value="<?php echo isset($id_batch) ?  $id_batch : ''; ?>">
-
-                    <div class="content-new-target">
-
-                        <!-- Input nuevo destinatario -->
-                        <div class="new-target">
-                            <input type="email" name="notifications_target" id="notifications_target" placeholder="Nuevo destinatario" required>
-                        </div>
-                        <!-- Boton de acciones -->
-                        <div class="acctions">
-
-                            <button type="submit" name="new_notification_target">
-                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
-                                    <path fill="#ffffff" d="M11 13H6q-.425 0-.712-.288T5 12q0-.425.288-.712T6 11h5V6q0-.425.288-.712T12 5q.425 0 .713.288T13 6v5h5q.425 0 .713.288T19 12q0 .425-.288.713T18 13h-5v5q0 .425-.288.713T12 19q-.425 0-.712-.288T11 18z" />
-                                </svg>
-                            </button>
-
-                        </div>
-                    </div>
-
-                </form>
+                <?php
+                if ($_SESSION['role'] == 'admin') {
+                    echo "
+                            <h3>Nuevo destinatarios</h3>
+                            <form class='input-new-target' method='post'>
+            
+                                <input type='hidden' name='id_batch' value=' ".(isset($id_batch) ?  $id_batch : '') ." '>
+                                
+                                    <div class='content-new-target'>
+            
+                                        <!-- Input nuevo destinatario -->
+                                        <div class='new-target'>
+                                            <input type='email' name='notifications_target' id='notifications_target' placeholder='Nuevo destinatario' required >
+                                        </div>
+                                        <!-- Boton de acciones -->
+                                        <div class='acctions'>
+                                            
+                                            <button type='submit' name='new_notification_target'>
+                                                <svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 24 24'>
+                                                    <path fill='#ffffff' d='M11 13H6q-.425 0-.712-.288T5 12q0-.425.288-.712T6 11h5V6q0-.425.288-.712T12 5q.425 0 .713.288T13 6v5h5q.425 0 .713.288T19 12q0 .425-.288.713T18 13h-5v5q0 .425-.288.713T12 19q-.425 0-.712-.288T11 18z' />
+                                                </svg>
+                                            </button>
+                
+                                        </div>
+                                    </div>
+                                
+            
+                            </form>
+                            ";
+                }
+                ?>
 
             </div>
             <div class="content-list-targets">
                 <h3>Lista de destinatarios actuales</h3>
-                
+
                 <div class="list-targets">
                     <?php
                     if ($_SERVER["REQUEST_METHOD"] == "POST" && isset($_POST['new_notification_target'])) {
-    
+
                         if (empty($_POST['notifications_target'])) {
                             $_SESSION['error-created'] = 'No puedes enviar este campo vacio';
                             return;
@@ -207,21 +211,26 @@ require_once '../app/controllers/batch/show_events.php';
                         echo '<p> No hay destinatarios</p>';
                         echo '<p> Agrega uno </p>';
                     } else {
-    
+
                         foreach ($_SESSION['notifications_target'] as $indice => $destinatario) {
                             echo "
-                                <div class='targets'>
-                                    <input name='' type='text' value='" . $destinatario . "'>
-                                    <button onclick='borrar(this)' data-indice='" . $indice . "'>
-                                        <a href='../app/controllers/batch/delete_notifications.php?indice=" . $indice . "'>X</a>
-                                       
-                                    </button>
+                            <div class='targets'>
+                                <input name='' type='text' value='" . $destinatario . "' disabled>
+                                ";
+                            if ($_SESSION['role'] == 'admin') {
+                                echo "
+                                        <button onclick='borrar(this)' data-indice='" . $indice . "'>
+                                            <a href='../app/controllers/batch/delete_notifications.php?indice=" . $indice . "'>X</a>
+                                        </button>
+                                    ";
+                            }
+                            echo "
                                 </div>
                             ";
                         }
                     }
                     ?>
-    
+
                 </div>
 
             </div>
@@ -257,33 +266,46 @@ require_once '../app/controllers/batch/show_events.php';
     <!-- -- EVENTOS --  -->
     <!-- Registrar evento -->
     <div class="container-events">
-        <form action="../app/controllers/batch/create_events.php" method="POST" class="register-events">
 
-            <div class="input">
-                <label for="modality">Modalidad</label>
-                <select name="modality" id="modality" required>
-                    <option value="">Seleccione</option>
-                    <option value="A">A = Presencial</option>
-                    <option value="V">V = Virtual</option>
-                </select>
-            </div>
+        <?php 
+        
+            if($_SESSION['role'] == 'admin'){
+                echo "
 
-            <div class="input">
-                <label for="training">Entrenamiento</label>
-                <select name="training" id="training" required>
-                    <option value="">Seleccione</option>
-                    <option value="2">2</option>
-                    <option value="6">6</option>
-                </select>
-            </div>
 
-            <div class="input">
-                <label for="seed_code">Código Semilla </label>
-                <input type="text" name="seed_code" id="seed_code" placeholder="Semilla" required>
-            </div>
+                <form action='../app/controllers/batch/create_events.php' method='POST' class='register-events'>
+        
+                    <div class='input'>
+                        <label for='modality'>Modalidad</label>
+                        <select name='modality' id='modality' required >
+                            <option value='>Seleccione</option>
+                            <option value='A'>A = Presencial</option>
+                            <option value='V'>V = Virtual</option>
+                        </select>
+                    </div>
+        
+                    <div class='input'>
+                        <label for='training'>Entrenamiento</label>
+                        <select name='training' id='training' required>
+                            <option value='>Seleccione</option>
+                            <option value='2'>2</option>
+                            <option value='6'>6</option>
+                        </select>
+                    </div>
+        
+                    <div class='input'>
+                        <label for='seed_code'>Código Semilla </label>
+                        <input type='text' name='seed_code' id='seed_code' placeholder='Semilla' required>
+                    </div>
+        
+                    <input type='submit' name='events' value='Añadir'>
+                </form>
 
-            <input type="submit" name="events" value="Añadir">
-        </form>
+
+                ";
+            }
+        ?>
+
 
         <!-- Lista de eventos -->
         <div class="events-sync">
@@ -349,7 +371,14 @@ require_once '../app/controllers/batch/show_events.php';
                             <th>Modalidad</th>
                             <th>Entrenamiento</th>
                             <th>Código de Semilla</th>
-                            <th>Acciones</th>
+                            <?php
+                                if($_SESSION['role'] == 'admin'){
+                                    echo "
+                                        <th>Acciones</th>
+                                    
+                                    ";
+                                }
+                            ?>
                         </tr>
                     </thead>
                     <tbody>
@@ -358,23 +387,37 @@ require_once '../app/controllers/batch/show_events.php';
                                 <td><?php echo $event['modality']; ?></td>
                                 <td><?php echo $event['training']; ?></td>
                                 <td><?php echo $event['seed_code']; ?></td>
-                                <!-- Agrega los enlaces de editar y borrar según sea necesario -->
-                                <td class="edit-icon">
-                                    <a onclick="handleEditEvent(event, '<?php echo $event['seed_code']; ?>', '<?php echo $event['modality']; ?>', '<?php echo $event['training']; ?>')" href="#">
-                                        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
-                                            <path fill="#39a900" d="M3 21v-4.25L16.2 3.575q.3-.275.663-.425t.762-.15q.4 0 .775.15t.65.45L20.425 5q.3.275.438.65T21 6.4q0 .4-.137.763t-.438.662L7.25 21zM17.6 7.8L19 6.4L17.6 5l-1.4 1.4z" />
-                                        </svg>
-                                    </a>
-                                    <a onclick="handleDeleteEvent('<?php echo $event['seed_code']; ?>')">
-                                        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
-                                            <g fill="none">
-                                                <path fill="#e11d48" fill-opacity=".25" d="m17.004 17.98l1.711-9.415c.117-.642.176-.963.013-1.049c-.162-.085-.394.145-.857.604l-.886.88L12 10L6.996 9l-.863-.865c-.465-.465-.697-.698-.86-.612c-.162.085-.104.408.014 1.054l1.71 9.402a.068.068 0 0 0 .018.036a7.05 7.05 0 0 0 9.97 0a.068.068 0 0 0 .019-.036" />
-                                                <ellipse cx="12" cy="7" stroke="#e11d48" stroke-linecap="round" stroke-width="1.2" rx="7" ry="3" />
-                                                <path stroke="#e11d48" stroke-linecap="round" stroke-width="1.2" d="m5 7l1.996 10.98a.068.068 0 0 0 .019.035v0a7.05 7.05 0 0 0 9.97 0v0a.068.068 0 0 0 .019-.036L19 7" />
-                                            </g>
-                                        </svg>
-                                    </a>
-                                </td>
+
+                                <?php
+                                    if($_SESSION['role'] == 'admin'){
+                                        echo "
+                                           
+
+                                        <!-- Agrega los enlaces de editar y borrar según sea necesario -->
+                                        <td class='edit-icon'>  
+
+                                            <a onclick=\"handleEditEvent('".$event['seed_code'] ."','". $event['modality'] ."', '".$event['training']."')\">
+                                            
+                                                <svg xmlns='http://www.w3.org/2000/svg' width='24' height='24' viewBox='0 0 24 24'>
+                                                    <path fill='#39a900' d='M3 21v-4.25L16.2 3.575q.3-.275.663-.425t.762-.15q.4 0 .775.15t.65.45L20.425 5q.3.275.438.65T21 6.4q0 .4-.137.763t-.438.662L7.25 21zM17.6 7.8L19 6.4L17.6 5l-1.4 1.4z' />
+                                                </svg>
+                                            </a>
+                                            <a onclick=\"handleDeleteEvent('".$event['seed_code']."')\">
+                                                <svg xmlns='http://www.w3.org/2000/svg' width='24' height='24' viewBox='0 0 24 24'>
+                                                    <g fill='none'>
+                                                        <path fill='#e11d48' fill-opacity='.25' d='m17.004 17.98l1.711-9.415c.117-.642.176-.963.013-1.049c-.162-.085-.394.145-.857.604l-.886.88L12 10L6.996 9l-.863-.865c-.465-.465-.697-.698-.86-.612c-.162.085-.104.408.014 1.054l1.71 9.402a.068.068 0 0 0 .018.036a7.05 7.05 0 0 0 9.97 0a.068.068 0 0 0 .019-.036' />
+                                                        <ellipse cx='12' cy='7' stroke='#e11d48' stroke-linecap='round' stroke-width='1.2' rx='7' ry='3' />
+                                                        <path stroke='#e11d48' stroke-linecap='round' stroke-width='1.2' d='m5 7l1.996 10.98a.068.068 0 0 0 .019.035v0a7.05 7.05 0 0 0 9.97 0v0a.068.068 0 0 0 .019-.036L19 7' />
+                                                    </g>
+                                                </svg>
+                                            </a>
+                                        </td>
+                                        
+                                        ";
+                                    }
+                                ?>
+
+
 
                             </tr>
                         <?php endforeach; ?>
